# Conflicts:
#	converters/pom.xml
#	core/pom.xml
#	demo/pom.xml
#	integration/extras/pom.xml
#	integration/pom.xml
#	integration/spring/pom.xml
#	integration/thymeleaf/pom.xml
#	integration/thymeleaf3/pom.xml
#	pom.xml

+++ core/src/main/java/com/github/resource4j/objects/providers/mutable/ResourceObjectIsNotBundleException.java
@@ -5,7 +5,7 @@ import com.github.resource4j.ResourceObjectException;
 
 public class ResourceObjectIsNotBundleException extends ResourceObjectException {
 
-    private static final int serialVersionUID = 1;
+    private static final long serialVersionUID = 1;
 
     public ResourceObjectIsNotBundleException(String name) {
         super(name);
